diff bug fix

+++ src/data/prepare_data.py
@@ -9,7 +9,7 @@ def prepare_dataset():
     df["unix"] = (df.created_at - pd.Timestamp("1970-01-01")) // pd.Timedelta('1s')
     df["banned"] = df.banned_at.isna().apply(lambda x: not x)
 
-    df["diff"] = df.banned.values - df.blocked.values
+    df["diff"] = df.banned.values.astype(int) - df.blocked.values.astype(int)
     df = df[df["diff"]==0]
 
     df = df[df["unix"] > BASE_STAMP][["created_at", "body", "blocked", "banned"]]
